fix: add splice test
